added actioncontext for dereference functions, moved with handler context to new context.py, added effector to both contexts, added deref call in default knowledge handler
