+++ src/koi_net/default_actions.py
@@ -1,11 +1,11 @@
-from .processor.handler_context import HandlerContext
+from .context import ActionContext
 from rid_lib.types import KoiNetNode
 from rid_lib.ext import Bundle
 from .effector import Effector
 
 
 @Effector.register_default_action(KoiNetNode)
-def dereference_koi_node(ctx: HandlerContext, rid: KoiNetNode) -> Bundle:
+def dereference_koi_node(ctx: ActionContext, rid: KoiNetNode) -> Bundle:
     if rid != ctx.identity.rid:
         return
