+++ src/koi_net/processor/knowledge_pipeline.py
@@ -21,7 +21,7 @@ from .knowledge_object import (
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from .handler_context import HandlerContext
+    from ..context import HandlerContext
 
 logger = logging.getLogger(__name__)
